requirements.txt
uvicorn==0.14.0
numpy>=1.21
pandas==1.3.5
python-dotenv==0.20.0
web3==5.28.0
tqdm
